Fix text-shrinking logic for pie-charts in KDChart.
The bug was that while calculating with the center we
then applied the logic to calc the different to the
whole size rather and just doubled the distance. That
could lead to text being shrinked way to much like
happened at bugs.kde.org bug 268481.

This is a follow-up for the fix of KDCH-484.

This bugfix was provided to you by the work done
on Calligra :-)



git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@43013 c66179b2-ff0d-0410-8aef-1b1a91e65cfa