Fix displaying of axis-labels if they would overlap.
Looks as that fix for bug 4179 was wrong and should never have been landed cause it completely breaks displaying of labels and ticks on the axis if the labels would overlap. In such cases we only displayed the first label+tick but completely ignored all following ones. Now we only skip the label that would overlap and then display only each e.g. 2th label while still drawing all ticks as expected. The proper fix for bug 4179 was landed later. So, no new regressions :)

For interest;
* the bug was introduced with r9202
* that whole commit just doesn't make any sense :-(
* this fix was provided to you by the work done on Calligra :-)



git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@43045 c66179b2-ff0d-0410-8aef-1b1a91e65cfa