Introduced possibility to fill the areas of a radar chart.
This introduces the KDChartRadarDiagram::setFillAlpha method
which allows to set an alpha-value for the color to fill the
areas of a radar chart with that color.

The default is 0.0 what means the areas are not filled with
any color just like before (what means the default behavior
is not changed). I introduced here a qreal rather then a
simple boolean-flag to still allow to draw the polylines
with the full brush-color while making it possible to use
the same color but more transparent for the areas to still
allow to see the other areas.

The logic introduced here is used in kchart@calligra to
implement filled radar charts.



git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@43252 c66179b2-ff0d-0410-8aef-1b1a91e65cfa