Rewrite CartesianAxis::Private::calculateMaximumSize().
Also disable clipping of axis painting because overpainting should, in
general, only happen when the size calculation is wrong.
The lengthwise size of the axis is not calculated by us so it's going to
be the same as the plane's, which causes clipping of the first and last
labels. To avoid having to calculate the exact axis length we just turn
off clipping.
Note that the old code tried to calculate axis length, but totally
failed. For example it returned 60 pixels when it was closer to 300.
Apparently that wasn't too much of a problem.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@45541 c66179b2-ff0d-0410-8aef-1b1a91e65cfa