Fix wrong legend colors (and other legend properties).
That was a regression introduced in r45612, to fix an issue in
KDCH-822. I think the real problem is that, with two-dimensional data,
an odd number of columns simply makes no sense. We should rather do a
sanity check and ignore the last column entirely. That still needs
to be done.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@46911 c66179b2-ff0d-0410-8aef-1b1a91e65cfa