Don't skip colors when working with two-dimensional data.
The AttributesModel must divide the column number by the dimension
count to obtain the dataset number and then get the the color number
dataset-number from the palette.

The addition of dataDimension to AttributesModel is binary
incompatible, so I have moved the data to a private class so it doesn't
happen again.
While changing so many lines I've also cleaned up some formatting and
verbose constructs, and deduplicated some copy-and-paste code in
compare() - note the new compareHeaderDataMaps().
Additionally, I noticed that every call to defaultHeaderData() asking
for a color would construct a throwaway palette with all its colors.
To fix that the palette is now being cached.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@46917 c66179b2-ff0d-0410-8aef-1b1a91e65cfa