Fix KDCH-917, "Axis floats away from the plane/grid".
First I have rewritten CartesionCoordinateTransformation because it was
involved here, and overcomplicated, and unnecessarily slow, and couldn't
do logarithmic back-transformation. This is all fixed now.
Note that the transformation is now more likely to return NaN because
a logarithmic scale cannot cross zero; if both positive and negative
values occur, one group must be transformed to NaN. The old code just
transformed the values to something that looked good in the logarithmic
plotter example...

The changed transformation broke CartesianGrid in logarithmic mode,
so instead of fixing its paint method I have rewritten that one using
TickIterator (originally written for CartesianAxis). Big win in terms
of code size and readability.

Finally, I implemented height-for-width support in CartesianGrid in
isometric scaling mode, which fixes the original bug. Doing this using
only QLayoutItem does not seem to work without resorting to a hack.
Due to that, in isometric scaling mode the the widget cannot be resized
by changing its height. Height can only be set indirectly via width.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@47047 c66179b2-ff0d-0410-8aef-1b1a91e65cfa