Change showZeroLabel() to showFirstTick() and enable it by default.
The intention seems to be enabling or disabling the label (and tick)
where the axes cross, which is not always at zero.
After the introduction of showZeroLabel(), the label was still enabled
by default on the X axis even though the default of showZeroLabel()
was false. I think it could really not be turned off at all on the X
axis.
Now both the X and Y first ticks are enabled by default and they both
can be turned off.
I've removed a special case that was only important with off by
default. Now I hope people just don't turn off the first label for
bar charts...

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@47418 c66179b2-ff0d-0410-8aef-1b1a91e65cfa