Make the tick label font slightly larger.
This affects both the scaled and the minimum size.
Tick labels are still significantly smaller than data value texts, but
I don't want to change the "aesthetics" of KD Chart too radically, and
this looks okay.
Data value texts are, with this change and the last change, ~42% larger
than tick labels. Before the difference was exactly 100%.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@47667 c66179b2-ff0d-0410-8aef-1b1a91e65cfa