Make the fixeDataCoordinateSpaceRelation thing more numerically stable.
Do not always multiply stuff onto the zoom values and geometry,
which accumulates errors and is somewhat brittle. Instead, store size
and zoom parameters from when fixed... was first enabled, never change
them, and use them to calculate the zoom parameters that keep diagram
size constant.
This is also necessary for the next commit.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@48245 c66179b2-ff0d-0410-8aef-1b1a91e65cfa