Fix a failing index range assertion by unscrewing the indexing.
qRound() was a worse choice than floor() here.
For example, with 1.01 indexesPerPixel() and 101 indexes / 100
CachePositions, on CachePosition 99 the previous algorithm iterated
from index 100 to 101. But 101 is invalid.
It also makes sense to stop where the next CachePosition starts, using
the same formula for start of current CachePosition and start of next
CachePositon, which marks the end of the current.
A side effect of this change is that the number of data points
returned per CachePosition varies, which is good for performance and
bad for visual consistency.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@48748 c66179b2-ff0d-0410-8aef-1b1a91e65cfa