Eliminate forwarding addLabel() methods.
Maybe they made more sense in 2.4 where the *DiagramType strangely
inherited *DiagramPrivate, which they don't anymore.
In 2.4 they  dropped the autoPositionNegative argument on the floor and
passed autoPositionPositive to the callee twice instead. That looked
wrong (like a copy & paste error - there was no comment about it), so
I "fixed" it in r44302. That caused a regression - for example, in line
diagrams the data point markers of negative values were always painted
at the very bottom of the diagram.
Now we just pass the right autoPositionFoo instead of "adjusting" them
in the forwarding addLabel() methods. Should look the same from
outside and it saves a couple of lines.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@48766 c66179b2-ff0d-0410-8aef-1b1a91e65cfa