Don't let markers and data value texts get into each other's way.
It's the same fix that I've previously applied to the AxisParameters
example, plus some straight code copying from over there because it
was less wrong (probably somebody fixed something in the meantime
over there).
That's the problem if you copy code around: You have to apply
fixes to all the copies.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@49804 c66179b2-ff0d-0410-8aef-1b1a91e65cfa