Fix geometry of Chart when drawing onto an auxiliary surface.
As tested in the DrawIntoWidget example.

It turns out that QLayout::activate() always tries to find a parent
QWidget and ask it for its QWidget::size(). And we don't set that size
in the paint call for auxiliary surfaces. Which makes sense because it
could trigger expensive and useless resizes of surrounding widgets when
the Chart is simultaneously used in a regular widget role.
QLayout::setGeometry() is our workaround. It seems to do roughly the
same thing as activate(), but takes the passed in geometry instead of
any QWidget's or parent's.
I am not sure why the x and y "starting position" (top left corner) of
the geometry must be moved to (0, 0) before passing it to the layout,
instead of passing on the rect as is. But it is necessary to get the
diagrams to appear where they are supposed to.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@50077 c66179b2-ff0d-0410-8aef-1b1a91e65cfa