Fix KDCH-963, wrong DataValueText positions in bar charts.
When the diagram is transposed, the default DataValueText position ends
up being different relative to the diagram's orientation. To fix that,
rotate the DataValueText position with the diagram.
This commit also moves the code to detect transposed diagrams from
CartesianAxis into AbstractDiagram::Private so it isn't duplicated
between the now multiple users.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@50139 c66179b2-ff0d-0410-8aef-1b1a91e65cfa