Fix KDCHDEV-810, missing data values.
No data was visible in the plot, it looked empty. The data was not set
correctly for 2-dimensional data and there was some code that did
nothing useful. So I've copied over the fixes for setting data from
the logarithmic plotter example. Regarding the invisible data, when
there are many points (like 10k) the data lines become faint. Maybe
some issue with antialiased one-pixel lines that are a fracton of a
pixel long? Needs further investigation.
For now, using "only" 1000 data points (instead of 100k) for about
1000 pixels width seems to be a good idea and makes the demo work.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@50143 c66179b2-ff0d-0410-8aef-1b1a91e65cfa