Fix KDCHDEV-34: Don't delete foreign items in the graphics scene.
Items created by KD Gantt are recreated in updateScene(), which is not
a big deal for API clients. But it was somewhat unexpected that
updateScene() also deleted "manually" inserted items, as one might do
to add arbitrary additional information to the view.
Now only items created by KD Gantt will be recreated in updateScene().

The new test fails before the change in GraphicsScene proper and
passes afterwards.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@50194 c66179b2-ff0d-0410-8aef-1b1a91e65cfa