Make isometric scaling work better and hopefully not crash anymore.
The biggest improvements are:
- Adjusting width instead of height when appropriate
- Removing the minimumSizeHint() hack

I've also tried to make size calculation more stable by using floating-
point and proper rounding wherever it makes sense. It didn't seem to
be a problem, though. Finally, the change in sizeHint() causes a more
suitable initial size of the whole window in the Plotter/Isometric
example. Without that, there was much more unused space in the initial
layout because it was too wide.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@50304 c66179b2-ff0d-0410-8aef-1b1a91e65cfa