Apply "customized label text" also when determining axis size.
This fixes another regression in the new axis code.
Previously, customized label text was only applied when painting for
real, so the estimate from calculateMaximumSize() could be smaller than
the real painted size.
A very nice side effect is that the hacks to make the estimated size
slightly larger "just in case" can be removed.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@50332 c66179b2-ff0d-0410-8aef-1b1a91e65cfa