Don't get stuck in CartesianAxis::Private::calculateMaximumSize().
Forward-port of r50489 from 2.5.1-release.

That happened in the CartesianGrid example before it could show
anything on screen, due to the TickIterator iterating forever without
reaching its end position.
The immediate cause was the data dimension spanning "from" NaN "to"
NaN. For a better fix, it would be good to catch the NaNs earlier.
In any case, it doesn't hurt to have this last line of defense.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@51282 c66179b2-ff0d-0410-8aef-1b1a91e65cfa