Fix KDCHDEV-836, which boils down to slow repaints due to weird values.
Extended forward-port of r50555 from kdchart-2.5.1-release.

QPainter does not like NaN (which were already caught) or infinite
values. It doesn't crash, but it slows down by a factor of about
40000 for affected lines, in the Plotter/Logarithmic example.
There are many other places that can hit the same slow path, but for
now just fix the one we've observed. It is also a rather common case.

The infinite values came from logarithmic trasformation of zero.
That's technically garbage input data, but a very very easy mistake
to make, so it's better to handle it gracefully.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@51283 c66179b2-ff0d-0410-8aef-1b1a91e65cfa