Unscrew how signals from models are handled.
Don't use queued connections.

Connect to the top of the model stack (AttributesModel) everywhere,
never to the second model from the top (the original source model).
This was already done that way in most places, just not everywhere.

This currently causes a bug: Each frame is drawn with the data
boundaries from the previous frame. Before this change, that only
happened under certain circumstances, like when the source model
was a proxy model. At least it's consistent now, and I'm working
on fixing it properly :)

Note that the AbstractItemView is still connected to the original
source model, not the AttributesModel, but the AbstractItemView
code really doesn't seem to do anything in KD Chart. It's just
a design error to base AbstractDiagram on it.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@51660 c66179b2-ff0d-0410-8aef-1b1a91e65cfa