Make sure that the data compressor receives updates before other code.
It is supposed to act like a a transparent cache, while the signals of
the underlying model are used elsewhere to get notified of changes.
So we have to enforce the invariant that whenever somebody receives a
notification from the model, the supposedly transparent compressor has
already updated itself. The "easiest" way to do this is making sure
that signal receivers are connected in the right order, i.e. the
compressor must always be the first to connect.

This fixes the bug that curves were drawn outside of the diagram
boundaries because the code that determines the boundaries and the
code that draws the values saw different data.
Now they both see up-to-date data.

A note regarding r51660: It looks like it wasn't the proxy model that
triggered the bug but using setModel() for a second time, regardless of
the model that was set.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@51672 c66179b2-ff0d-0410-8aef-1b1a91e65cfa