Match the right Legend constructor with Legend(AbstractDiagram*).
It used to invoke the QWidget * overload which didn't do what a user
could expect. Now, the second overload is a closer match, and it does
the right thing and calls setDiagram() on the diagram.
The alternative fix, using qobject_cast in the one-argument QWidget*
constructor and setDiagram() if the QObject is-a
KDChart::AbstractDiagram, was considered and rejected. Similar effect
but more of a hack.

Source incompatible change, but it should only change things for the
better. This was an unnecessary pitfall.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@52415 c66179b2-ff0d-0410-8aef-1b1a91e65cfa