Fix wrong pie chart label protrusion calculation, KDCH-995.
When slice labels would be painted outside the designated pie area, the
pie size is decreased, thus moving the labels (placed relative to the
pie) inwards.
This broke when the painter was transformed, which happens only with
external painting, i.e. when calling
Chart::paint( QPainter* painter, const QRect& target ).
Most of KD Chart assumes that the painter isn't transformed, which
works surprisingly well as long as one doesn't compare untransformed
with transformed coordinates. This kind of comparison happens in pie
chart with labels. So ensure the same transformed-ness for both
rectangles that are compared.

This is not the big fix I was hoping for. I went for less regression
potential and less work here... Not sure if a better fix is even
possible, due to the externally visible methods involved.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@53205 c66179b2-ff0d-0410-8aef-1b1a91e65cfa