Fix TextLayoutItem::intersects() when both items have the same angle.
That was the fast path, which didn't work, so just remove it.
Optimize the slow path a little instead. We don't need to translate
both polygons / regions, and intersects() should be faster than
actually calculating the intersection.

This fixes that non-horizontal axis labels were sometimes wrongly found
to overlap and shortened or decimated aka thinned out accordingly. This
caused little quirks that were not bad enough for somebody to bother
and file a bug report. It still made a bad impression. The Axis/Chart
example was affected, for example.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@55188 c66179b2-ff0d-0410-8aef-1b1a91e65cfa