Fix unit prefixes or suffixes being added twice.
This is a fix for r51741.
headerDataLabels() already have unit prefixes and suffixes added, which
is fine, we just need to know it and not do it again. Fortunately,
CartesianAxis::customizeLabel() is supposed to be called as the very
last step just before the text is displayed (*fixes already added),
which we can easily do here.

Also apply the change from r51741 *and* the fix to it in
CartesianAxis::calculateMaximumSize(), so that size calculation is
hopefully synced again with the actual painting code.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@55189 c66179b2-ff0d-0410-8aef-1b1a91e65cfa