Deprecate the totally broken CartesianAxis::setTitleSpace().
I think it mostly existed to "fill the gap" in the Axis/Chart example,
top axis. The gap was there because shortened labels were wrongly used
due to the just fixed bug in TextLayoutItem::intersects().
Since there might be no other "good" use for this feature, I am not
fixing the following issues:
- Semantics - what does the titleSpace value *mean*?
  (fixing this requires adding a new method that does it right)
- (related) make sure not to divide by zero
- Layout - adjust the size reported by calculateMaximumSize() according
  to spacing

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@55191 c66179b2-ff0d-0410-8aef-1b1a91e65cfa