Make sure to update legends when changing diagram type, KDCHDEV-841.
This is a rather hacky fix, it relies on changes to the diagrams
(like setModel() here) triggering relayout in the legends. It would be
better to ensure that legends are always correctly rebuilt when
attaching them to a different diagram. In fact the code tries hard
enough to do that, in several places, but sometimes it stil doesn't
work.
I suspect that legend layout updates stop at some sort of "update
barrier", like setting a geometry equal to the current geometry is
caught early and the main part of the work is prevented.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@55233 c66179b2-ff0d-0410-8aef-1b1a91e65cfa