Reduce hackiness of, and fix legend layout.
This seems to be another regression introduced by not rebuilding the
Chart's layout at every resize. The fix here is calling legend->show()
in Chart::addLegend(), then we don't need as much hackery to paint and
lay out an "invisible" widget. Unfortunately, I couldn't remove all
the hacks because that causes regressions e.g. with external painting
like in the DrawIntoPainter example or KD Reports.
It is probably possible to clean up legend layout much more, but the
current state seems to work in all cases we care about which I will
declare good enough at this point.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@56125 c66179b2-ff0d-0410-8aef-1b1a91e65cfa