Re-add Legend::resizeLayout(), a reimplemented virtual.
I somehow missed that it was reimplemented and not added in Legend.
This partially reverts r56125. The most important part (even though
smallest) part there was calling show() on the legend. It looked like
resizeLayout() wasn't required anymore, but it turns out that removing
it sometimes causes legends that have the right external size but
wrong internal layout. That makes sense, since Legend is *required*
by the AbstractAreaWidget API contract to reimplement resizeLayout()
and adjust its internal layout (if any) there.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@57312 c66179b2-ff0d-0410-8aef-1b1a91e65cfa