Align axes in their grid layouts, which matters with axis sharing.
This fixes an important layout bug for AIM (via e-mail, not in Jira)
and, strangely, even seems to fix laterally cut off axes (until a
window resize triggers a more complete layout run) from the same
"torture test" code from AIM as well.
Maybe not so strangely if the cut-off was also due to position off
by a few pixels.

With aligned(*) diagrams and therefore axes, the widest axis determines
the grid cell's width, and the narrower ones are not supposed to kind
of hang out somewhere in their cell. They should always be as close to
the diagram as possible in their grid cell.

(*) I forgot how exactly to get that configuration, but I can *see* the
layout tree using a custom layout tree printing debug helper.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@57316 c66179b2-ff0d-0410-8aef-1b1a91e65cfa