Center axes in the longitudinal direction.
This somehow fixes the bug that axis titles were not centered
longitudinally. Ticks and tick labels were fine because their painting
uses the coordinate plane's coordinate translation, which is mostly
independent of any QLayout business.

This worked automagically with no alignment flags at all (presumably
defaulting to AlignHCenter | AlignVCenter). Since r57316 we set the
alignment flags, so we need to se both of them - for the longitudinal
direction we need centered, instead of nothing (int 0).

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@58272 c66179b2-ff0d-0410-8aef-1b1a91e65cfa