CartesianAxis: Adapt ~all geometry when relevant attributes change.
This is inefficient and unfortunate (and a bit ugly), but probably
required for correctness.

Fixes a bug for AIM (reported by e-mail) where adding annotations to
an axis with axis title would not change the axis title's position to
move away from the then longer tick labels.

Afer fixing that, I applied the same fix to similar code.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@60485 c66179b2-ff0d-0410-8aef-1b1a91e65cfa