KDCH-1026: fix check in ModelDataCache.
The source model has already been updated (rows removed) here, the
cache not yet. It does not make sense to check that the range is in
the now smaller range of the input model. It does make sense to check
that the range is in the not yet shrunk cache data.

git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@60979 c66179b2-ff0d-0410-8aef-1b1a91e65cfa